Hide WalkedEvent from EventProcessor
When paging data WalkedEvent will be required
When processing an event only the target event type will be exposed. This simplifies accessing the data and hides the cast internally. The cast is safe enough, though there is nothing at compile time to guarantee the IPager in use is providing pages of the correct event type.